backend/requirements.txt
alpha-vantage==2.1.1
appnope==0.1.0
asn1crypto==1.0.1
astroid==2.1.0
attrs==19.2.0
autopep8==1.4.3
backcall==0.1.0
bleach==3.1.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
convertdate==2.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
cycler==0.10.0
Cython==0.29.14
decorator==4.4.0
defusedxml==0.6.0
Django==2.1.5
django-cors-headers==2.4.0
django-filter==2.1.0
django-hitcount==1.3.0
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.1
entrypoints==0.3

holidays==0.9.11
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.4
itypes==1.1.0
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
lunardate==0.2.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.2
openapi-codec==1.3.2
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
plotly==4.2.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
pycodestyle==2.4.0
pycparser==2.19
Pygments==2.4.2
pylint==2.2.2
PyMeeus==0.3.6
pyparsing==2.4.2
pyrsistent==0.15.4
pyspellchecker==0.5.0
pystan==2.19.1.1
python-dateutil==2.8.0
pytz==2018.9
pyzmq==18.1.0
qtconsole==4.5.5
redis==3.3.11
requests==2.22.0
retrying==1.3.3
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
setuptools-git==1.2
simplejson==3.16.0
six==1.12.0
sklearn==0.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.6
virtualenv==16.7.7
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.0
zipp==0.6.0
drf-yasg